Listen on inotify event file close
Earlier we listened on even IN_CREATE. IN_CREATE does not guarantee that
the apport crash has fully been written. This could be the reason why
apport kept reporting us that the report was corrupt. By using
IN_CLOSE_WRITE, we ensure to act only on events where the file was
opened to be written, and then was closed.

Thanks: Laurent Bigonville
Closes: #750989