Ony honor IN_CREATE. Also act only for .crash files
As per the inotify man page:

IN_CREATE (+)
File/directory created in watched directory  (e.g., open(2)  O_CREAT,
mkdir(2),  link(2),  sym‐link(2), bind(2) on a UNIX domain socket).

So if apport writes to a file (which would mean open(), then we'll act
on it.

The tests for it looked good.